+++ Plugins/ServerSorter/ServerSorter.plugin.js
@@ -6,7 +6,7 @@ class ServerSorter {
 	getName() { return "ServerSorter"; }
 	getShortName() { return "ServerSorter"; }
 	getDescription() { return "Adds server sorting abilities to Discord. Support Server: bit.ly/ZeresServer"; }
-	getVersion() { return "0.2.5"; }
+	getVersion() { return "0.3.0"; }
 	getAuthor() { return "Zerebos"; }
 	
 	load() {}
@@ -29,7 +29,6 @@ class ServerSorter {
 	stop() {
 		$(document).add("*").off(this.getShortName());
 		$(".guild-sorter").remove();
-		$("#sort-options").remove();
 		BdApi.clearCSS(this.getShortName());
 	}
 
@@ -39,47 +38,35 @@ class ServerSorter {
 		"#sort-options.open { pointer-events:initial;opacity:1;transition: 300ms cubic-bezier(.2,0,0,1); transform-origin: 0 0; transform: translateY(0px);}");
 		this.SortedGuildStore = InternalUtilities.WebpackModules.findByUniqueProperties(['getSortedGuilds']);
 		
-		this.sorter = $('<div class="guild guild-sorter" id="bd-pub-li" style="height: 20px; margin-bottom:10px;"><div class="guild-inner" style="height: 20px; border-radius: 4px;"><a><div id="bd-pub-button" class="sort-button" style="line-height: 20px; font-size: 12px;">Sort</div></a></div></div>');
-		this.options = $(`<div id="sort-options" class="contextMenu-uoJTbz theme-dark" style="left: 10px;">
-		<div class="itemGroup-oViAgA">
-		<div class="item-1XYaYf" data-sort="name" data-reverse="false"><span>Alphabetically</span><div class="hint-3TJykr">A > Z</div></div>
-		<div class="item-1XYaYf" data-sort="name" data-reverse="true"><span>Reverse Alphabetically</span><div class="hint-3TJykr">Z > A</div></div>
-		</div>
-		<div class="itemGroup-oViAgA">
-		<div class="item-1XYaYf" data-sort="joinedAt" data-reverse="true"><span>Newest Joined</span><div class="hint-3TJykr">New</div></div>
-		<div class="item-1XYaYf" data-sort="joinedAt" data-reverse="false"><span>Oldest Joined</span><div class="hint-3TJykr">Old</div></div>
-		</div>
-		<div class="itemGroup-oViAgA">
-		<div class="item-1XYaYf" data-sort="id" data-reverse="true"><span>Newest Created</span><div class="hint-3TJykr"></div></div>
-		<div class="item-1XYaYf" data-sort="id" data-reverse="false"><span>Oldest Created</span><div class="hint-3TJykr"></div></div>
-		</div>
-		<div class="itemGroup-oViAgA">
-		<div class="item-1XYaYf" data-sort="id" data-reset-sort="true"><span>Reset</span></div></div>
-		</div>`);
-		this.options.find('span').attr('style',"width: 110px;display: inline-block;overflow: hidden;text-overflow: ellipsis;");
-		this.options.find('.item-1XYaYf').on("click." + this.getShortName(), (e) => {
-			var item = $(e.currentTarget);
-			if (item.data("sort")) this.doSort(item.data("sort"), item.data("reverse"), item.data("reset-sort"));
-		}); 
-		this.options.appendTo('#app-mount');
+		let sortButton = $('<div class="guild guild-sorter" id="bd-pub-li" style="height: 20px; margin-bottom:10px;"><div class="guild-inner" style="height: 20px; border-radius: 4px;"><a><div id="bd-pub-button" class="sort-button" style="line-height: 20px; font-size: 12px;">Sort</div></a></div></div>');
 
-		this.sortButton = $(this.sorter.find('.sort-button')[0]);
-		this.sortButton.on("click." + this.getShortName(), () => {
-			this.options.css("top", 10 + this.sorter.offset().top + this.sorter.height() + "px");
-			this.options.toggleClass('open');
+		let contextMenu = new PluginContextMenu.Menu().addItems(
+			new PluginContextMenu.ItemGroup().addItems(
+				new PluginContextMenu.TextItem("Alphabetically", {hint: "A > Z", callback: () => {this.doSort("name", false);}}),
+				new PluginContextMenu.TextItem("Reverse Alphabetical", {hint: "Z > A", callback: () => {this.doSort("name", true);}})
+			),
+			new PluginContextMenu.ItemGroup().addItems(
+				new PluginContextMenu.TextItem("Newest Joined", {hint: "New", callback: () => {this.doSort("joinedAt", false);}}),
+				new PluginContextMenu.TextItem("Oldest Joined", {hint: "Old", callback: () => {this.doSort("joinedAt", true);}})
+			),
+			new PluginContextMenu.ItemGroup().addItems(
+				new PluginContextMenu.TextItem("Newest Created", {callback: () => {this.doSort("id", false);}}),
+				new PluginContextMenu.TextItem("Oldest Created", {callback: () => {this.doSort("id", true);}})
+			),
+			new PluginContextMenu.ItemGroup().addItems(
+				new PluginContextMenu.TextItem("Reset", {danger: true, callback: () => {this.doSort("id", false, true);}})
+			)
+		);
+
+		sortButton.find('.sort-button').on("click." + this.getShortName(), (e) => {
+			contextMenu.show(e.clientX, e.clientY);
 		});
-		this.sorter.insertBefore($('.guild-separator'));
-		$(window).on("click." + this.getShortName(), (e) => {
-			if (!((e.pageY > this.sortButton.offset().top && e.pageY < this.sortButton.offset().top + this.sortButton.height()) &&
-				(e.pageX > this.sortButton.offset().left && e.pageX < this.sortButton.offset().left + this.sortButton.width()))) {
-				this.options.removeClass('open');
-			}
-		});
+		sortButton.insertBefore($('.dms + .guild-separator'));
 		PluginUtilities.showToast(this.getName() + " " + this.getVersion() + " has started.");
 	}
 
 	getGuilds() {
-		return $('div.guild:has(div[draggable="true"])');
+		return $('div.guild:has(div[draggable="true"]):not(#server-search)');
 	}
 	
 	getGuildData(guild) {